move _render_template to where it is singularly used
